test(zero-server): expand request benchmarks
